@@ -3,8 +3,9 @@
   "description": "A short description of what your package does",
   "license": "MIT",
   "require": {
-    "php": ">=7.1",
-    "ext-curl": "*"
+    "php": ">=7.4",
+    "ext-curl": "*",
+    "ext-json": "*"
   },
   "autoload": {
     "psr-4": {
